Extract custom assertion into tools
