+++ Tests/FoundationExtrasTests/Persistence/Storage/MemoryStorageTests.swift
@@ -7,21 +7,6 @@
 import FoundationExtras
 import XCTest
 
-public func XCTAssertThrowsError<T>(
-    _ expression: @autoclosure () async throws -> T,
-    _ message: @autoclosure () -> String = "",
-    file: StaticString = #filePath,
-    line: UInt = #line,
-    _ errorHandler: (_ error: any Error) -> Void = { _ in }
-) async {
-    do {
-        _ = try await expression()
-        XCTFail(message())
-    } catch {
-        errorHandler(error)
-    }
-}
-
 // MARK: - MemoryStorageTests
 
 final class MemoryStorageTests: XCTestCase {
